NPC now move on predefined paths

+++ src/PaooGame/GameObjects/Entity.java
@@ -7,7 +7,7 @@ public abstract class Entity extends GameObject {
 
     protected int speed;    // Speed
     // Direction variables | last input sets sign based on direction
-    protected byte xSign, ySign;
+    protected int xSign, ySign;
 
 
     public Entity(int x, int y) {
@@ -18,10 +18,10 @@ public abstract class Entity extends GameObject {
     public int getSpeed() {
         return speed;
     }
-    public byte getXSign() {
+    public int getXSign() {
         return xSign;
     }
-    public byte getYSign() {
+    public int getYSign() {
         return ySign;
     }
 }
